LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/uievents/mouse/mouse_boundary_events_after_removing_last_over_element-expected.txt


FAIL After removing the last over element, redundant mouseenter events should not be fired on the ancestors assert_equals: expected "mouseover@div#parent, mouseout@div#parent, mouseleave@div#parent, mouseleave@div#grandparent, mouseover@body, mousemove@body" but got "mouseout@div#child, mouseleave@div#child, mouseover@div#parent, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mousemove@div#parent, mouseout@div#parent, mouseleave@div#parent, mouseleave@div#grandparent, mouseover@body, mousemove@body"
FAIL After removing the root element in the shadow under the cursor, mouseleave events should be targeted outside the shadow, but redundant mouseenter events should not be fired assert_equals: expected "mouseover@div#shadowHost, mouseout@div#shadowHost, mouseleave@div#shadowHost, mouseleave@div#containerOfShadowHost, mouseover@body, mousemove@body" but got "mouseout@div#divInShadow, mouseleave@div#divInShadow, mouseover@div#shadowHost, mouseenter@body, mouseenter@div#containerOfShadowHost, mouseenter@div#shadowHost, mousemove@div#shadowHost, mouseout@div#shadowHost, mouseleave@div#shadowHost, mouseleave@div#containerOfShadowHost, mouseover@body, mousemove@body"
